openswe cli: streaming service
